[PluginAPI] Build the test plugins rather than compiling them
They used whatever clang was on PATH, which Bazel's sandbox does not provide.

The python bindings also needed a header dep and a dead variable removed, or
the Bazel compiler tree would not build.
