+++ compiler/src/iree/compiler/PluginAPI/test/CMakeLists.txt
@@ -3,27 +3,37 @@ iree_add_all_subdirs()
 # The plugin mechanism is not yet supported on Windows.
 if (NOT WIN32)
 
+# Built rather than compiled inside a RUN line, so the tests need no compiler of
+# their own. Plain C, and deliberately not renamed: they are never registered.
+foreach(_PLUGIN "valid_plugin" "invalid_plugin" "version_mismatch_plugin")
+  add_library(iree_compiler_PluginAPI_test_${_PLUGIN} MODULE
+    "plugins/${_PLUGIN}.c")
+  set_target_properties(iree_compiler_PluginAPI_test_${_PLUGIN} PROPERTIES
+    POSITION_INDEPENDENT_CODE ON)
+endforeach()
+
 iree_lit_test_suite(
   NAME
     lit
   SRCS
     "invalid_options.txt"
-    "invalid_plugin.c"
-    "valid_plugin.c"
-    "version_mismatch.c"
+    "invalid_plugin.txt"
+    "valid_plugin.txt"
+    "version_mismatch.txt"
   TOOLS
     FileCheck
     iree-compile
     iree-opt
 )
 
-# These build a plugin, so give lit the compiler this build used rather than
-# leaving %host_cc to whatever PATH offers.
 iree_package_ns(_PACKAGE_NS)
 string(REPLACE "::" "/" _PACKAGE_PATH ${_PACKAGE_NS})
-foreach(_TEST_SRC "invalid_plugin.c" "valid_plugin.c" "version_mismatch.c")
-  set_property(TEST "${_PACKAGE_PATH}/${_TEST_SRC}.test"
-    APPEND PROPERTY ENVIRONMENT "IREE_HOST_CC=${CMAKE_C_COMPILER}"
+foreach(_TEST_SRC "invalid_plugin.txt" "valid_plugin.txt" "version_mismatch.txt")
+  set_property(TEST "${_PACKAGE_PATH}/${_TEST_SRC}.test" APPEND PROPERTY
+    ENVIRONMENT
+      "IREE_TEST_VALID_PLUGIN=$<TARGET_FILE:iree_compiler_PluginAPI_test_valid_plugin>"
+      "IREE_TEST_INVALID_PLUGIN=$<TARGET_FILE:iree_compiler_PluginAPI_test_invalid_plugin>"
+      "IREE_TEST_VERSION_MISMATCH_PLUGIN=$<TARGET_FILE:iree_compiler_PluginAPI_test_version_mismatch_plugin>"
   )
 endforeach()
