+++ compiler/src/iree/compiler/PluginAPI/test/plugins/invalid_plugin.c
@@ -1,0 +1,11 @@
+// Copyright 2026 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+// A shared library carrying no plugin entry point.
+
+#include <stdbool.h>
+
+bool some_function(void) { return false; }
