+++ compiler/src/iree/compiler/PluginAPI/test/valid_plugin.txt
@@ -1,0 +1,4 @@
+// A plugin built as a module rather than compiled here, so the test needs no
+// compiler of its own.
+
+// RUN: iree-opt --iree-load-plugin=$IREE_TEST_VALID_PLUGIN --help
